Post voting.
Post @OneToMany votes.
user @OneToMany votes.
Made a PostVote class that has a PostVoteId, a Post that it belongs to, a User that voted, and the type of post it is indicated by an int, 1 or -1. PostVoteId is neat because it shows you don't need to make a column in the database and that the id is embeddable in PostVote. PostVote has methods that declare a vote(), and methods that checks if a vote isUpVote or isDownVote and who voteBelongsTo()